+++ features/foundation-info/src/main/java/com/example/foundation_info/data/api/foundation/FoundationApi.kt
@@ -8,6 +8,6 @@ interface FoundationApi {
 
     @GET("foundation")
     suspend fun getFoundation(
-        @Query("id") id: Int
+        @Query("id") id: Long
     ): FoundationResponse
 }
